Added detailsFramework:CreateAuraScrollBox(parent, name, data, onAuraRemoveCallback, options)
Creates a scrollbox to display auras.

Data: is a table<spellId, boolean> where the boolean if true indicates the spell is tracked by its Name (false tracks by the spellId).

onAuraRemoveCallback: calls this function when the user click in the X button to remove the aura.

options: a table of type df_aurascrollbox_options